feat : 간단한 디렉토리 리팩토링
