+++ src/components/pages/stock/Trade.tsx
@@ -2,8 +2,8 @@
 import { getInvestors, getStockName } from '@/actions/stock/stock'
 import { getAccessToken } from '@/actions/tokens'
 import { wonInfoAPI } from '@/actions/wallet'
-import SocketTradeModal from '@/components/ui/SockTradeModal'
-import TradeModal from '@/components/ui/TradeModal'
+import SocketTradeModal from '@/components/pages/stock/SockTradeModal'
+import TradeModal from '@/components/pages/stock/TradeModal'
 import { toast } from '@/components/ui/use-toast'
 import { staticStockType } from '@/types/Stock'
 import { socketStockCode } from '@/utils/socketStockCode'
