--- src/components/ui/TradePassword.tsx
@@ -1,3 +1,0 @@
-export default function TradePassword() {
-  return <div>dfkslfjs</div>
-}
